### 2024-08-12 4cb6f13
Add NS note to guide popup
