### 2024-08-12 #7097 @see24
Hi @kthyng, thanks for your review. I just noticed that one of the figures isn't rendering properly in the PDF so I will look into that.

Potential reviewers:

- Jamie M. Kass; jamiemkass; worked on wallace R package which I think is similar.
- Cristina Ronquillo; cRonFer; worked on OCCUR, a fairly complex Shiny App that might be similar
- Stephanie Pennington; stephpenn1; In JOSS database, has Shiny experience
- Zeke Marshall; ZekeMarshall; In JOSS database, has Shiny experience
- Maximilian Hesselbarth; mhesselbarth; In JOSS database, lots of spatial R experience
- Nicolas Casajus; ahasverus; In JOSS database, lots of spatial R experience
