### 2024-08-12 cab20eb
fix fig 1 in paper
