+++ inst/joss_paper/paper.md
@@ -30,7 +30,7 @@ Climate change vulnerability assessments (CCVAs) are tools increasingly adopted 
 
 The NatureServe CCVI algorithm combines several components that contribute to a species' vulnerability to climate change, including exposure to changes in temperature and moisture (Section A), indirect exposure to other changes brought on by climate change (e.g. sea level rise; Section B), and the species' traits that either make it sensitive to climate change or allow it to adapt to climate change (Section C; \autoref{fig:NS-alg}). In addition, it optionally incorporates the results of documented or modeled responses to climate change [Section D\; @young2012; @young2016]. The resulting CCVI can have one of five index values, 'Less Vulnerable', 'Moderately Vulnerable', 'Highly Vulnerable', 'Extremely Vulnerable' or 'Insufficient Evidence'. Exposure is assessed by determining the proportion of the species' range that falls into 6 classes of temperature and moisture change, which is used to determine an exposure multiplier. Indirect exposure, sensitivity and adaptive capacity (Sections B and C), are assessed by scoring 23 vulnerability factors on a scale from ‘neutral’ (score: 0) to ‘greatly increases vulnerability’ (score: 3). Factors that cannot be scored can be left blank and do not contribute to the total score, but if fewer than 13 factors are scored the final index value will be returned as 'Insufficient Evidence'. The indirect exposure, sensitivity and adaptive capacity scores are then multiplied by the exposure multiplier and summed giving a total score. An index value for Sections A-C is then determined by applying a set of thresholds to the total score. If documented or modeled responses to climate change are included they are also scored on the scale from ‘neutral’ (score: 0) to ‘greatly increases vulnerability’ (score: 3) and then summed to give a score for Section D, which is then converted to an index value based on a threshold. The index values for Sections A-C and Section D are then combined into a final index value using a table that gives more weight to the sensitivity and adaptive capacity sections (\autoref{fig:NS-alg}).
 
-![Algorithm for calculating the NatureServe Climate Change Vulnerability Index.\label{fig:NS-alg}](NS_ccvi_alg_diagram.png)
+![Algorithm for calculating the NatureServe Climate Change Vulnerability Index.\label{fig:NS-alg}](NS_CCVI_alg_diagram.png)
 
 The NatureServe CCVI tool is available as an Excel workbook that users can fill in based on (a priori) spatial analyses, available research or expert opinion. While this Excel tool is easy to access and straightforward to use, it requires users to possess technical Geographical Information System (GIS) skills and software, that is often proprietary, to perform the spatial components of the analysis. This challenges the implementation of reproducible analyses, which are increasingly required for credible scientific processes [@munafò2017]. In addition, the Excel-based version requires repeated assessments to determine the effects of uncertainty associated with choice of Global Climate Models (GCMs) or emission scenarios.
 
@@ -42,7 +42,7 @@ Additional novel features of `ccviR` include functionality to incorporate multip
 
 # Statement of need
 
[…]
