### 2024-08-16 #7097 @kthyng
@see24 Please trim down your paper a bit. It shouldn't be more than 1000 words ideally — you could shorten the wording or move some sections to the docs.
