+++ inst/joss_paper/paper.Rmd
@@ -26,17 +26,17 @@ journal: JOSS
 
 # Summary
 
-Climate change vulnerability assessments (CCVAs) are tools increasingly adopted to rank species' vulnerability to the threat of climate change [@pacifici2015]. Common CCVA approaches include trait-based, correlative models (e.g. species distribution models), mechanistic models (e.g. mechanistic niche models), or a combination of these approaches [@foden2019; @pacifici2015]. CCVAs can be used to inform extinction risk assessments by identifying the mechanisms and magnitude of impacts [@foden2019]. One popular trait-based CCVA tool is the [NatureServe Climate Change Vulnerability Index](https://www.natureserve.org/conservation-tools/climate-change-vulnerability-index) (CCVI), which is a rapid assessment tool designed to rank species' vulnerability to climate change and to highlight which factors contribute to the climate change vulnerability of individual species or groups of taxa [@young2016; @young2015]. This information can inform conservation decision making and identify actions to increase species' resilience to climate change.
+Climate change vulnerability assessments (CCVAs) are tools increasingly adopted to rank species' vulnerability to the threat of climate change [@pacifici2015]. Common CCVA approaches include trait-based, correlative models (e.g. species distribution models), mechanistic models (e.g. mechanistic niche models), or a combination of these approaches [@foden2019; @pacifici2015]. CCVAs can be used to inform extinction risk assessments by identifying the mechanisms and magnitude of impacts [@foden2019]. One popular trait-based CCVA tool is the [NatureServe Climate Change Vulnerability Index](https://www.natureserve.org/conservation-tools/climate-change-vulnerability-index) (CCVI), which is a rapid assessment tool designed to rank species' vulnerability to climate change and to highlight which factors contribute to the vulnerability of individual species or groups[@young2016; @young2015]. This information can inform conservation decisions and identify actions to increase species' resilience to climate change.
 
-The NatureServe CCVI algorithm combines several components that contribute to a species' vulnerability to climate change, including exposure to changes in temperature and moisture (Section A), indirect exposure to other changes brought on by climate change (e.g. sea level rise; Section B), and the species' traits that either make it sensitive to climate change or allow it to adapt to climate change (Section C; \autoref{fig:NS-alg}). In addition, it optionally incorporates the results of documented or modeled responses to climate change [Section D\; @young2012; @young2016]. The resulting CCVI can have one of five index values, 'Less Vulnerable', 'Moderately Vulnerable', 'Highly Vulnerable', 'Extremely Vulnerable' or 'Insufficient Evidence'. Exposure is assessed by determining the proportion of the species' range that falls into 6 classes of temperature and moisture change, which is used to determine an exposure multiplier. Indirect exposure, sensitivity and adaptive capacity (Sections B and C), are assessed by scoring 23 vulnerability factors on a scale from ‘neutral’ (score: 0) to ‘greatly increases vulnerability’ (score: 3). Factors that cannot be scored can be left blank and do not contribute to the total score, but if fewer than 13 factors are scored the final index value will be returned as 'Insufficient Evidence'. The indirect exposure, sensitivity and adaptive capacity scores are then multiplied by the exposure multiplier and summed giving a total score. An index value for Sections A-C is then determined by applying a set of thresholds to the total score. If documented or modeled responses to climate change are included they are also scored on the scale from ‘neutral’ (score: 0) to ‘greatly increases vulnerability’ (score: 3) and then summed to give a score for Section D, which is then converted to an index value based on a threshold. The index values for Sections A-C and Section D are then combined into a final index value using a table that gives more weight to the sensitivity and adaptive capacity sections (\autoref{fig:NS-alg}).
[…]
