### 2024-08-16 #7097 @see24
> @see24 Please trim down your paper a bit. It shouldn't be more than 1000 words ideally — you could shorten the wording or move some sections to the docs.

Ok I made an attempt to trim down the wording a little but it didn't achieve much. One option I see would be to remove the whole 2nd paragraph since there is a similar section in the app_details_vignette.Rmd. But would that be too much? I want to make sure people can still understand the context of how the index works.
@kthyng or @mengqi-z I would appreciate your guidance!
