### 2024-08-16 #7097 @kthyng
@see24 I think it would be good to keep some of the second paragraph for context, like you said, but I think you could cut it off midway, somewhere around the sentence starting with "Factors that cannot be stored...". After do that you can probably call it good, you don't have to be strictly below 1000 words if you aren't at that point, but we appreciate your efforts to do so.
