### 2024-08-19 #7097 @mengqi-z
> Hey @mengqi-z, unfortunately I am lacking the time at the moment. But maybe @lmjoconnor or @rcornf are available?

@mhesselbarth - Thank you for your reply and recommendations!

@lmjoconnor @rcornf - Would you be interested in reviewing this submission for JOSS? The review process is checklist-driven and conducted directly within JOSS GitHub issues, following these guidelines here: https://joss.readthedocs.io/en/latest/review_criteria.html
