### 2024-08-19 #7097 @sebdalgarno
hi @mengqi-z - I can take on reviewing this JOSS submission
