+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   app_vignette: app_vignette.html
   data_prep_vignette: data_prep_vignette.html
   package_vignette: package_vignette.html
-last_built: 2024-08-16T20:10Z
+last_built: 2024-08-20T12:58Z
 urls:
   reference: https://landscitech.github.io/ccviR/reference
   article: https://landscitech.github.io/ccviR/articles
