### 2024-08-20 #7097 @see24
@kthyng I edited paragraph 2 I didn't want to just cut it off half way because then it wouldn't mention how section D is scored so it would be a bit confusing. I tried to simplify the second half though. Let me know if you have any more suggestions!
