+++ inst/joss_paper/paper.Rmd
@@ -28,13 +28,13 @@ journal: JOSS
 
 Climate change vulnerability assessments (CCVAs) are tools increasingly adopted to rank species' vulnerability to the threat of climate change [@pacifici2015]. Common CCVA approaches include trait-based, correlative models (e.g. species distribution models), mechanistic models (e.g. mechanistic niche models), or a combination of these approaches [@foden2019; @pacifici2015]. CCVAs can be used to inform extinction risk assessments by identifying the mechanisms and magnitude of impacts [@foden2019]. One popular trait-based CCVA tool is the [NatureServe Climate Change Vulnerability Index](https://www.natureserve.org/conservation-tools/climate-change-vulnerability-index) (CCVI), which is a rapid assessment tool that ranks species' vulnerability to climate change and highlights factors contributing to increased vulnerability[@young2016; @young2015]. Outputs of the NatureServe CCVI can inform conservation decisions and identify actions to increase species' resilience to climate change.
 
-The NatureServe CCVI combines several components that contribute to a species' vulnerability to climate change, including exposure to changes in temperature and moisture (Section A), indirect exposure to other changes brought on by climate change (e.g. sea level rise; Section B), and traits related to species' sensitivity or ability to adapt to climate change (Section C; \autoref{fig:NS-alg}). It also optionally incorporates the results of documented or modeled responses to climate change [Section D\; @young2012; @young2016]. The resulting CCVI ranks include: 'Less Vulnerable', 'Moderately Vulnerable', 'Highly Vulnerable', 'Extremely Vulnerable' or 'Insufficient Evidence'. Exposure is assessed by determining the proportion of the species' range that falls into six classes of temperature and moisture change, which is used to assign an exposure multiplier. Sections B, C and optionally D are each scored on a scale from ‘neutral’ (score: 0) to ‘greatly increases vulnerability’ (score: 3). The indirect exposure, sensitivity and adaptive capacity scores are multiplied by the exposure multiplier and summed. The final CCVI rank is a result of combining all scores through a variety of operations including weighting sensitivity and adpative capacity scores. [comment: i think in the process of reducing the following text, a lot is lost that affects comprehension, Consider deleting the following sentences and replacinng with my suggestions, above ] Where applicable the score is multiplied by an exposure multiplier and then scores are summed separately for Sections B and C and Section D. Thresholds are applied to these scores to determine an index value and then the two index values are combined into a final index value (i.e., rank) using a table that gives more weight to the sensitivity and adaptive capacity sections (\autoref{fig:NS-alg}).
[…]
