### 2024-08-22 bea652a
remove extra refs in bib, add doi
