+++ inst/joss_paper/paper.bib
@@ -28,7 +28,7 @@
 	date = {2012},
 	journal = {Wildlife Conservation in a Changing Climate},
 	pages = {129--152},
-	url = {https://www.degruyter.com/document/doi/10.7208/9780226074641-007/pdf?licenseType=restricted}
+	url = {https://www.degruyter.com/document/doi/10.7208/9780226074641-007/pdf?licenseType=restricted},
 	note = {ISBN: 9780226074641}
 }
 
