### 2024-08-23 #7097 @mengqi-z
👋 @skiptoniam @quang-huynh @rhenkin @jamiemkass - would any of you be interested in reviewing this JOSS submission? The review process is checklist-driven and conducted directly within JOSS GitHub issues, following these guidelines here: https://joss.readthedocs.io/en/latest/review_criteria.html
