### 2024-08-27 #7150 @editorialbot
[REVIEW]: ccviR: an R package and Shiny app to implement the NatureServe Climate Change Vulnerability Index
**Submitting author:** @see24 (<a href="http://orcid.org/0000-0001-9644-5343">Sarah Endicott</a>)
**Repository:** https://github.com/LandSciTech/ccviR
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.1
