### 2024-08-29 LandSciTech/ccviR#177 @sebdalgarno
R-CMD check test error
openjournals/joss-reviews#7150

I'm getting some test errors in `test-run_ccvi_app.R`

Can you check on your end?
