### 2024-08-29 LandSciTech/ccviR#178 @sebdalgarno
fix no visible binding for global variable R-CMD check notes
https://github.com/openjournals/joss-reviews/issues/7150

this might not technically be required for publication - but worth addressing I think

need to fix in dplyr and tidyr functions
data masking and tidy selection sections in [programming with dplyr](https://dplyr.tidyverse.org/articles/programming.html#tidy-selection) vignette should be useful
