### 2024-08-29 LandSciTech/ccviR#179 @sebdalgarno
kableExtra 'namespace not imported from' R-CMD check note
https://github.com/openjournals/joss-reviews/issues/7150

might not be technically required for publication but perhaps worth addressing

`library(kableExtra)` is called in results_report.Rmd - are any kableExtra functions used?
