### 2024-08-30 LandSciTech/ccviR#177 @see24
Yes I get an error on my end as well. I have had a lot of trouble with shinytest and have since given up on it. I would like to leave the file in case I am able to get it working again. For now I just have some manual tests that I run for the app (See #151) . All the non-shiny package functions have testthat tests though.
