### 2024-08-30 LandSciTech/ccviR#179 @see24
It seems that R-CMD check doesn't look at files in inst when it decides if the package is used. I added a reference to a kableExtra function in the R folder so it should be imported from now.
