### 2024-08-30 5133ee1
fixes #178 avoid global variable notes
