### 2024-08-30 LandSciTech/ccviR#178 @see24
Done. Just as a heads up the article you linked to isn't very helpful for this situation, maybe you meant this: [Using dplyr in packages](https://dplyr.tidyverse.org/articles/in-packages.html#data-masking-and-tidy-selection-notes)?
