### 2024-08-30 7d716ef
add test for html exposure tables
