### 2024-08-30 82ea919
try to get shinytest2 working again
