### 2024-08-30 42c452f
shinytest2 app will start, spatial not running
