+++ tests/testthat/test-app.R
@@ -1,2 +1,0 @@
-# shinytest2::test_app("app", check_setup = FALSE)
-test_dir(test_path("app"))
