+++ tests/testthat/app/app.R
@@ -1,3 +1,0 @@
-options(shiny.testmode = TRUE)
-
-ccviR:::ccvi_app()
