### 2024-09-02 LandSciTech/ccviR#177 @sebdalgarno
would it be possible to remove (or comment out) the stop call in this test so that all tests pass?
```
test_that("save works especially for spatial and range change and all related inputs", {
  skip_on_cran()
  skip_on_ci()
  stop("add test")
})
```
