### 2024-09-03 c2a4657
skip app tests until they are ready
