### 2024-09-03 ad47b32
add stringi to suggests for tests
