### 2024-10-08 #7150 @mengqi-z
@sebdalgarno Awesome! Thank you so much for taking the time and effort to review this submission – your contribution is greatly appreciated!

JOSS is always looking for more reviewers like you. I noticed you're not yet in our reviewer database, and if you're interested in reviewing more exciting papers in the future, we'd love to have you join! You can easily add your profile here https://reviewers.joss.theoj.org/reviewers/new, so that other editors can find you for papers that match your expertise!
