### 2024-10-15 b021a8a
minor UI edits based on Ilona's comments in #181
