+++ R/ccvi_app.R
@@ -202,7 +202,7 @@ ccvi_app <- function(testmode_in, ...){
                   to evaluate select questions in Section C that have a spatial
                   component. If provided, the range change raster(s) will be used to
                   evalaute questions about the modeled response to climate change
-                  in Section D."),
+                  in Section D. Required datasets are indicated with", labelMandatory("a")),
                 get_file_ui("clim_var_dir", "Folder location of prepared climate data",
                             type = "dir", mandatory = TRUE, spinner = TRUE),
                 verbatimTextOutput("clim_var_error", ),
@@ -339,17 +339,17 @@ ccvi_app <- function(testmode_in, ...){
                 id = "secB",
                 h3("Section B: Indirect Exposure to Climate Change"),
                 h5("Evaluate for assessment area under consideration"),
-                check_comment_ui("B1", "1) Exposure to sea level rise:",
+                check_comment_ui("B1", "B1) Exposure to sea level rise:",
                                  choiceNames = valueNms,
                                  choiceValues = valueOpts),
-                check_comment_ui("B2a", "2a) Distribution relative to natural barriers",
+                check_comment_ui("B2a", "B2 a) Distribution relative to natural barriers",
                                  choiceNames = valueNms,
                                  choiceValues = valueOpts),
-                check_comment_ui("B2b", "2b) Distribution relative to anthropogenic barriers",
+                check_comment_ui("B2b", "B2 b) Distribution relative to anthropogenic barriers",
                                  choiceNames = valueNms,
                                  choiceValues = valueOpts),
 
-                check_comment_ui("B3", "  3) Predicted impact of land use changes resulting from human responses to climate change",
+                check_comment_ui("B3", "B3) Predicted impact of land use changes resulting from human responses to climate change",
                                  choiceNames = valueNms[2:4],
                                  choiceValues = valueOpts[2:4])
               ),
@@ -363,33 +363,33 @@ ccvi_app <- function(testmode_in, ...){
               div(
                 id = "secC",
                 h3("Section C: Sensitivity and Adaptive Capacity"),
-                check_comment_ui("C1", "1) Dispersal and movements",
+                check_comment_ui("C1", "C1) Dispersal and movements",
                                  choiceNames = valueNms,
                                  choiceValues = valueOpts),
 
-                strong("2b) Predicted sensitivity to changes in precipitation, hydrology, or moisture regime:"),
+                strong("C2 b) Predicted sensitivity to changes in precipitation, hydrology, or moisture regime:"),
 
-                check_comment_ui("C2bii", "ii) physiological hydrological niche.",
+                check_comment_ui("C2bii", "C2 b ii) physiological hydrological niche.",
                                  choiceNames = valueNms,
                                  choiceValues = valueOpts),
 
-                check_comment_ui("C2c", "2c) Dependence on a specific disturbance regime likely to be impacted by climate change.",
+                check_comment_ui("C2c", "C2 c) Dependence on a specific disturbance regime likely to be impacted by climate change.",
                                  choiceNames = valueNms[2:4],
                                  choiceValues = valueOpts[2:4]),
-                check_comment_ui("C2d", "2d) Dependence on ice, ice-edge, or snow-cover habitats.",
+                check_comment_ui("C2d", "C2 d) Dependence on ice, ice-edge, or snow-cover habitats.",
                                  choiceNames = valueNms,
                                  choiceValues = valueOpts),
 
-                check_comment_ui("C3", "3) Restriction to uncommon landscape/geological features or derivatives.",
+                check_comment_ui("C3", "C3) Restriction to uncommon landscape/geological features or derivatives.",
[…]
