### 2024-10-23 7006d64
fix naming or range change rasters with scenario names
