### 2024-10-29 #7150 @stephpenn1
Hi @see24 and @mengqi-z, thank you for your patience and thank you for the opportunity to review this. Please note I reviewed this with Ciara Donegan (@ciara-donegan), an early career scientist interested in gaining reviewing experience.

Overall, we agreed the Shiny app is interesting and well organized. The functionality is generally intuitive (probably more intuitive for those more familiar with the topic) and does it's intended job. We have a few points for feedback:

_App:_
- The app description and paper state that you don't need GIS skills, but step 1 seems to require at least _some_ familiarity?(e.g. if you want to use a region other than the demo one)
- In the _Vulnerability Questions_ and _Spatial Vulnerability Questions_ section, there could be some text to let the user know you may select multiple options. Similar to what is stated in the [vignette](https://github.com/LandSciTech/ccviR/blob/master/vignettes/app_vignette.Rmd#L104).
- This is not critical to the functionality, merely an optional suggestion for future versions: The report rendering can take awhile and to the user, seems like it's frozen since the progress bar stops. The progress bar you have is a great solution to this, but could have more progress increments taken from the R console outputs.

_Paper:_
- The paper is well written and easy to follow. I do think it could use a small section on using the app, similar to the github page instructions _Launching the App_ https://github.com/LandSciTech/ccviR?tab=readme-ov-file#launching-the-app
