### 2024-11-04 8fbb9c3
don't say no GIS skills, add launching the app to paper
