+++ inst/joss_paper/paper.Rmd
@@ -34,15 +34,29 @@ The NatureServe CCVI combines several components that contribute to a species' v
 
 The NatureServe CCVI tool is an Excel workbook that users complete using available scientific information, expert opinion, and spatial analyses. While the tool is straightforward to use, it requires users to possess technical Geographical Information System (GIS) skills and access to  software (often proprietary) to perform the spatial analyses. This challenges the implementation of reproducible analyses, which are increasingly required for credible scientific processes [@munafò2017]. In addition, the Excel-based version requires repeated assessments to assess uncertainty associated with choice of Global Climate Models or emission scenarios.
 
-To improve accessibility and reproducibility of the NatureServe CCVI we developed `ccviR`, an R package and Shiny app that implements the original CCVI in an easy to use Graphical User Interface (GUI). `ccviR` applies the original scoring algorithm to assess vulnerability and includes the original Monte Carlo uncertainty analysis. A major advance of `ccviR` is the functionality to perform spatial analyses internally, using R, such that GIS skills and costly software are not required. Users are only required to specify the locations of spatial data files on their computer, using the Shiny app or R code, and `ccviR` automates all spatial analyses. Examples of these automated spatial processes include overlays of the species' range with exposure categories, historical climate regime, and modeled changes in the species' range. The `ccviR` Shiny app presents the results of the spatial analyses using interactive maps, that allow for user-directed validation of the spatial analyses and opportunities to modify the resulting factor scores, if desired. Additional interactive plots allow users to explore what factors are driving the vulnerability of a species (\autoref{fig:ccviR-app}). Another significant addition is a second Shiny app that allows users to classify independently-supplied climate data sets into the exposure categories used by the index based on the median and half the interquartile range. This makes the methods used to determine exposure classes explicit and repeatable for different climate data sets or different regions. Finally,`ccviR` includes functionality to assess the consequences of uncertainty associated with choice of emission scenarios and/or global circulation models on CCVI outcomes. 
+To improve accessibility and reproducibility of the NatureServe CCVI we developed `ccviR`, an R package and Shiny app that implements the original CCVI in an easy to use Graphical User Interface (GUI). `ccviR` applies the original scoring algorithm to assess vulnerability and includes the original Monte Carlo uncertainty analysis. A major advance of `ccviR` is the functionality to perform spatial analyses internally, using R, such that advanced GIS skills and costly software are not required. Users are only required to specify the locations of spatial data files on their computer, using the Shiny app or R code, and `ccviR` automates all spatial analyses. Examples of these automated spatial processes include overlays of the species' range with exposure categories, historical climate regime, and modeled changes in the species' range. The `ccviR` Shiny app presents the results of the spatial analyses using interactive maps, that allow for user-directed validation of the spatial analyses and opportunities to modify the resulting factor scores, if desired. Additional interactive plots allow users to explore what factors are driving the vulnerability of a species (\autoref{fig:ccviR-app}). Another significant addition is a second Shiny app that allows users to classify independently-supplied climate data sets into the exposure categories used by the index based on the median and half the interquartile range. This makes the methods used to determine exposure classes explicit and repeatable for different climate data sets or different regions. Finally,`ccviR` includes functionality to assess the consequences of uncertainty associated with choice of emission scenarios and/or global circulation models on CCVI outcomes. 
 
[…]
