+++ README.Rmd
@@ -53,7 +53,7 @@ run_ccvi_app()
 ccviR uses the same vulnerability factors and scoring algorithm as the original NatureServe Excel spreadsheet. The index values, scores and Monte Carlo uncertainty analysis produced are the same.
 
 ### New Features
-- Spatial analyses are included in the package so no GIS skills are needed 
+- Spatial analyses are included in the package so only minimal GIS skills are needed 
 - Uses climate data from the whole species range rather than the range in the assessment area to score thermal and hydrological niche factors
 - Simultaneously calculates the index for multiple scenarios such as, emissions scenarios, time horizons or GCMs
 - A function and Shiny app to classify new climate data sets into exposure categories 
