### 2024-11-04 #7150 @see24
@mengqi-z and @stephpenn1 I have addressed Steph's comments as follows:
- The app description and paper state that you don't need GIS skills, but step 1 seems to require at least some familiarity?(e.g. if you want to use a region other than the demo one)
  - I have modified the text to make it more clear that some GIS skills are in fact needed [Commit](https://github.com/LandSciTech/ccviR/commit/8fbb9c3b96c716e1368de555e19e0a7b593d6095)
- In the Vulnerability Questions and Spatial Vulnerability Questions section, there could be some text to let the user know you may select multiple options. Similar to what is stated in the [vignette](https://github.com/LandSciTech/ccviR/blob/master/vignettes/app_vignette.Rmd#L104).
  - Yes, I agree
 and have added this. [Commit](https://github.com/LandSciTech/ccviR/commit/c9ee660ec26cc4e0cb75323e716975b17fafde86)
- This is not critical to the functionality, merely an optional suggestion for future versions: The report rendering can take awhile and to the user, seems like it's frozen since the progress bar stops. The progress bar you have is a great solution to this, but could have more progress increments taken from the R console outputs.
  - Yes I agree, we are actually planning to rejig the report so that it outputs using html rather than pdf since LaTEX dependencies can be a pain for users. We will implement your suggestion when we do that. [Issue](https://github.com/LandSciTech/ccviR/issues/163)

- The paper is well written and easy to follow. I do think it could use a small section on using the app, similar to the github page instructions Launching the App https://github.com/LandSciTech/ccviR?tab=readme-ov-file#launching-the-app
  - I have added this to the paper [Commit](https://github.com/LandSciTech/ccviR/commit/8fbb9c3b96c716e1368de555e19e0a7b593d6095)

Thanks for all of your efforts in reviewing!
