### 2024-11-04 #7150 @mengqi-z
@see24 @stephpenn1 - Thank you both!

@see24 - I’ll proofread the paper and ensure it meets JOSS standards, then follow up with you on the next steps. Thanks!
