### 2024-11-08 LandSciTech/ccviR#186 @mengqi-z
JOSS paper proofread
Hi @see24,

Thank you for your submission to JOSS! Before we can proceed with acceptance, a few more issues need to be addressed: https://github.com/openjournals/joss-reviews/issues/7150

It appears that your latest updates for the paper are in `inst/joss_paper/paper.Rmd`, while `inst/joss_paper/paper.md` is outdated. Since the JOSS PDF will be rendered from `paper.md`, the old version will be used in its current state. Could you please update `paper.md` to ensure it reflects your latest revisions?

A few additional minor comments regarding the JOSS paper, based on the latest [PDF](https://github.com/openjournals/joss-reviews/issues/7150#issuecomment-2455716576) rendered from `paper.md`:

1. On line 71, please format ccviR in code style (i.e., `ccviR`), rather than with quotation marks.
2. For the citation to Young et al., 2012, please consider updating the URL to: https://www.degruyter.com/document/doi/10.7208/9780226074641-007.
3. The paper is very well written overall. My only suggestion is for the Summary section, which currently feels quite lengthy. As this section primarily serves to provide a high-level overview of the software's functionality and purpose, it may make sense to divide it into three sections: Summary, State of the Field, and Functionality/Example. However, I’ll leave this decision up to you.
