+++ inst/joss_paper/paper.Rmd
@@ -26,11 +26,15 @@ journal: JOSS
 
 # Summary
 
+## Background
+
 Climate change vulnerability assessments (CCVAs) are tools increasingly adopted to rank species' vulnerability to the threat of climate change [@pacifici2015]. Common CCVA approaches include trait-based, correlative models (e.g. species distribution models), mechanistic models (e.g. mechanistic niche models), or a combination of these approaches [@foden2019; @pacifici2015]. CCVAs can be used to inform extinction risk assessments by identifying the mechanisms and magnitude of impacts [@foden2019]. One popular trait-based CCVA tool is the [NatureServe Climate Change Vulnerability Index](https://www.natureserve.org/conservation-tools/climate-change-vulnerability-index) (CCVI), which is a rapid assessment tool that ranks species' vulnerability to climate change and highlights factors contributing to increased vulnerability[@young2016; @young2015]. Outputs of the NatureServe CCVI can inform conservation decisions and identify actions to increase species' resilience to climate change.
+
+![Algorithm for calculating the NatureServe Climate Change Vulnerability Index.\label{fig:NS-alg}](NS_CCVI_alg_diagram.png)
 
 The NatureServe CCVI combines several components that contribute to a species' vulnerability to climate change, including exposure to changes in temperature and moisture (Section A), indirect exposure to other changes brought on by climate change (e.g. sea level rise; Section B), and traits related to species' sensitivity or ability to adapt to climate change (Section C; \autoref{fig:NS-alg}). It also optionally incorporates the results of documented or modeled responses to climate change [Section D\; @young2012; @young2016]. The resulting CCVI ranks include: 'Less Vulnerable', 'Moderately Vulnerable', 'Highly Vulnerable', 'Extremely Vulnerable' or 'Insufficient Evidence'. Exposure is assessed by determining the proportion of the species' range that falls into six classes of temperature and moisture change, which is used to assign an exposure multiplier. Sections B, C and optionally D are each scored on a scale from ‘neutral’ (score: 0) to ‘greatly increases vulnerability’ (score: 3). The indirect exposure, sensitivity and adaptive capacity scores are multiplied by the exposure multiplier and summed. The final CCVI value is a result of applying thresholds to the total scores and balancing the sensitivity and adaptive capacity rank with the documented or modeled response rank (\autoref{fig:NS-alg}).
 
-![Algorithm for calculating the NatureServe Climate Change Vulnerability Index.\label{fig:NS-alg}](NS_CCVI_alg_diagram.png)
+## Functionality
 
 The NatureServe CCVI tool is an Excel workbook that users complete using available scientific information, expert opinion, and spatial analyses. While the tool is straightforward to use, it requires users to possess technical Geographical Information System (GIS) skills and access to  software (often proprietary) to perform the spatial analyses. This challenges the implementation of reproducible analyses, which are increasingly required for credible scientific processes [@munafò2017]. In addition, the Excel-based version requires repeated assessments to assess uncertainty associated with choice of Global Climate Models or emission scenarios.
 
@@ -40,7 +44,7 @@ The `ccviR` Shiny app can be launched from an R session with one line of code an
 
 ![Interactive visualizations in the ccviR app allow validation and interpretation of results for a) the calculation of the temperature exposure multiplier and b) the scores for each vulnerability factor used to calculate the index.\label{fig:ccviR-app}](app_visuals.png)
 
-## Launching the app
+## Example
 
 The code below will open the app in your default browser with an example data set available. 
 
@@ -56,7 +60,7 @@ run_ccvi_app()
 
 # Statement of need
 
[…]
