### 2024-11-08 LandSciTech/ccviR#186 @see24
Thanks @mengqi-z!

I have made the changes you suggested and knit the paper.md so it is now up to date. For 3 I created sub-headings with in Summary section to create more structure because I know a Summary section is required but I didn't feel like the first paragraphs really give a full summary. Hopefully you think that is an improvement. Let me know!
