### 2024-11-08 #7150 @mengqi-z
Hi @see24 - I’ve completed the proofreading for the paper, so we’re ready for the next steps:

- [ ] Verify that all authors, affiliations, and ORCIDs are correct.
- [ ] Make a release of the software with the latest changes from the review and **post the version number here**. This is the version that will be used in the JOSS paper.
- [ ] Archive the released software on Zenodo or a similar service (e.g., figshare, an institutional repository).
- [ ] Make sure that the archive (e.g., on Zenodo) has the correct metadata. This includes the title (which should match the paper title) and the author list (ensure the list is correct and does not include people who only made minor contributions). You may also add the authors' ORCIDs.
- [ ] Please **post the DOI of the archived version here**.

Once these steps are complete, I’ll proceed with accepting the submission.
