### 2024-11-15 10dd1b4
add doi to README
