### 2024-11-15 #7150 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
