### 2024-11-15 #7150 @mengqi-z
@see24 - Thank you! I’ve recommended the acceptance of your paper. At this stage, the Editor-in-Chief (EiC) will take over and conduct final checks before proceeding with publication. We appreciate your submission to JOSS! Great work!
