### 2024-11-15 #7150 @see24
@kthyng  fixed in this [commit](https://github.com/LandSciTech/ccviR/commit/e030c8167d212f1f973552c4c86bb53d981b18ee)

Do I need to update the version and make a new release?
