### 2024-11-15 #7150 @kthyng
@see24 Glad you had a great experience! If you want to join JOSS as a reviewer, please sign up here: https://reviewers.joss.theoj.org/
